requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
requests==2.31.0
openpyxl==3.1.2
pytest==7.4.2
pytest-flask==1.2.0
httpie==2.6.0
googlemaps==4.10.0
